Chain panic hooks
Rather than trying to recreate the default panic hook for when
the panic does not match the expected signature, the hook ignoring
an expected panic can chain to the previously installed hook.

The unit tests demonstrate how to set each hook only once with
std::sync::Once. This is needed because Rust's test runner does
not provide a way to initalize global state for tests, and the
tests are ran in parallel by default.