Renamed JoinHandle -> Handle
The name should not be easily confused with std::thread's JoinHandle.
It should also not be very long, because it has to be used in explicit
annotations for the return type of spawn(). I didn't have better ideas
than Handle.

Also provided as_ and into_ accessors to the encapsulated thread
join handle.