Redesign for a more ergonomic API
No more explicit type coercion, yay!
The JoinHandle'y type got a properly descriptive name,
ControlledJoinHandle, because nobody cares when they don't have to
type it.

spawn() is now a method on Context. To call the constructor or the
From<thread::Builder> impl of Context, it only has to be
turbofish-qualified with one parameter for the expected panic type.