Remove workarounds for resolve_trait_on_defaulted_unit lint
The tests, especially doctests, should not rely on #[allow] when
a completely lint-free alternative is available.

The example of problematic use is left as a doctest in the examples
to Context::spawn(), with more drill-down explanation on the issue
including a compile-fail example that will also serve as a tripwire
for when feature(never_type) lands in stable Rust, and the final
lint-free variant.