CircleCI: Goodbye Machine, it was nice to know you!
It turns out --security-opt seccomp=unconfined is available out of the
box with vanilla CircleCI docker executors. Sheesh.

The Codecov upload runs as a separate job because it needs a
curl-equipped container which ragnaroek/kcov isn't.

Make kcov produce separate coverage reports per executable.
If allowed to store multiple reports in the same directory in successive
runs, kcov creates a misleading header in the coverage report index page,
as if it referred to the last executable run.